+++ packages/fighting-design/list/src/props.ts
@@ -37,4 +37,4 @@ export const Props = {
 
 export type ListPropsType = ExtractPropTypes<typeof Props>
 
-export const listPropsKey: InjectionKey<ListPropsType> = Symbol('list-props')
+export const LIST_PROPS_KEY: InjectionKey<ListPropsType> = Symbol('list-props-key')