+++ packages/fighting-design/loading/index.ts
@@ -1,11 +1,17 @@
-import Loading from './src/loading.vue'
+import type Loading from './src/loading.vue'
 
-import { install } from '../_utils'
+import { vLoading } from './src/directive'
 
-export const FLoading = install(Loading)
+import { install, installDirective } from '../_utils'
+
+export const FLoadingDirective = installDirective(vLoading, 'loading')
+
+export const FLoading = install({
+  directive: FLoadingDirective
+})
 
 export type LoadingInstance = InstanceType<typeof Loading>
 
 export * from './src/interface.d'
 
-export default Loading
+export default FLoading