+++ packages/fighting-design/loading/src/loading.vue
@@ -3,40 +3,51 @@
   import { Props } from './props'
   import { FSvgIcon } from '../../svg-icon'
   import { FIconLoadingAVue } from '../../_svg'
+  import type { ClassListInterface } from '../../_interface'
   import type { ComputedRef, CSSProperties } from 'vue'
-  import type { HandleMouseEventInterface } from '../../_interface'
   import type { LoadingPropsType } from './props'
 
   const prop: LoadingPropsType = defineProps(Props)
-
-  // 点击
-  const handleClick: HandleMouseEventInterface = (evt: MouseEvent): void => {
-    prop.closeEnd && prop.closeEnd(evt)
-  }
 
   /**
    * 样式列表
    */
   const styleList: ComputedRef<CSSProperties> = computed((): CSSProperties => {
-    const { background, opacity } = prop
+    const { background, fontColor } = prop
 
     return {
       background,
-      opacity
+      color: fontColor
     } as const
   })
+
+  /**
+   * 类名列表
+   */
+  const classList: ComputedRef<ClassListInterface> = computed(
+    (): ClassListInterface => {
+      const { fullscreen = false } = prop
+
+      return [
+        'f-loading',
+        {
+          'f-loading__fullscreen': fullscreen
+        }
+      ] as const
+    }
+  )
 </script>
 
 <template>
-  <div v-if="show" class="f-loading" :style="styleList" @click="handleClick">
+  <div v-if="visible" :class="classList" :style="styleList">
     <f-svg-icon :size="20" class="f-loading__animation" :icon="icon">
       <slot name="icon">
         <f-icon-loading-a-vue />
       </slot>
     </f-svg-icon>
 
-    <span class="f-loading__title" :style="{ fontSize, color: fontColor }">
-      {{ text || ' 玩命加载中...' }}
+    <span v-if="text" class="f-loading__title">
+      {{ text }}
     </span>
   </div>
 </template>